Adding wrapping code to Results instead
