+++ pymeasure/experiment/results.py
@@ -26,6 +26,9 @@ import logging
 
 import os
 import re
+import sys
+from copy import deepcopy
+from importlib.machinery import SourceFileLoader
 from datetime import datetime
 
 import pandas as pd
@@ -136,6 +139,39 @@ class Results(object):
                     f.write(self.header())
                     f.write(self.labels())
             self._data = None
+
+    def __getstate__(self):
+        # Get all information needed to reconstruct procedure
+        self._parameters = self.procedure.parameter_values()
+        self._class = self.procedure.__class__.__name__
+        module = sys.modules[self.procedure.__module__]
+        self._package = module.__package__
+        self._module = module.__name__
+        self._file = module.__file__
+
+        state = self.__dict__.copy()
+        del state['procedure']
+        del state['procedure_class']
+        return state
+
+    def __setstate__(self, state):
+        self.__dict__.update(state)
+
+        # Restore the procedure
+        module = SourceFileLoader(self._module, self._file).load_module()
+        cls = getattr(module, self._class)
+
+        self.procedure = cls()
+        self.procedure.set_parameters(self._parameters)
+        self.procedure.refresh_parameters()
+
+        self.procedure_class = cls
+
+        del self._parameters
+        del self._class
+        del self._package
+        del self._module
+        del self._file
 
     def header(self):
         """ Returns a text header to accompany a datafile so that the procedure
